requirements.txt
alembic==1.0.9
Babel==2.6.0
blinker==1.4
Click==7.0
cycler==0.10.0
dominate==2.3.5
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Excel==0.0.7
flask-ldap3-login==0.9.16
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kiwisolver==1.1.0
ldap3==2.6
lml==0.0.9
Mako==1.0.9
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyecharts-jupyter-installer==0.0.3
pyexcel==0.5.13
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.17
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
texttable==1.6.1
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
